Fixed crash importing a heightmap larger than the landscape #jira UE-35054 #rb Jack.Porter
[CL 3097574 by Gareth Martin in 4.13 branch]

+++ Engine/Source/Editor/LandscapeEditor/Private/LandscapeEditorDetailCustomization_NewLandscape.cpp
@@ -871,8 +871,8 @@ FReply FLandscapeEditorDetailCustomization_NewLandscape::OnCreateButtonClicked()
 			const TArray<uint16>& ImportData = LandscapeEdMode->UISettings->GetImportLandscapeData();
 			if (ImportData.Num() != 0)
 			{
-				const int32 OffsetX = (SizeX - ImportSizeX) / 2;
-				const int32 OffsetY = (SizeY - ImportSizeY) / 2;
+				const int32 OffsetX = (int32)(SizeX - ImportSizeX) / 2;
+				const int32 OffsetY = (int32)(SizeY - ImportSizeY) / 2;
 
 				// Heightmap
 				Data = LandscapeEditorUtils::ExpandData(ImportData,
